addressed review comments: feedback extraction boundary
